# server/core-api/models/model_linked_user.go
/*
 * Swagger solo - OpenAPI 3.0
 *
 * solo API
 *
 * API version: 1.0.0
 * Contact: apiteam@solo.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package models

type LinkedUser struct {
	Model

	UserID uint
	User   *User `json:"user,omitempty"`
	// account status
	Permissions string `json:"permissions,omitempty"`
}