# server/core-api/models/model_user.go
/*
 * Swagger solo - OpenAPI 3.0
 *
 * solo API
 *
 * API version: 1.0.0
 * Contact: apiteam@solo.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package models

type User struct {
	Model

	Username string `json:"username,omitempty" gorm:"unique"`

	Email string `json:"email,omitempty" gorm:"unique"`

	Password string `json:"password,omitempty"`
	// User Status
	UserStatus int32 `json:"userStatus,omitempty"`
	// Accounts that belong to user
	Networks []Network `json:"networks,omitempty"`

	Verified bool `json:"verified,omitempty"`

	Avatar string `json:"avatar,omitempty"`
}